fix bug

+++ mmcv/ops/csrc/common/cuda/active_rotated_filter_cuda_kernel.cuh
@@ -12,10 +12,9 @@
 
 template <typename scalar_t>
 __global__ void ARF_forward_cuda_kernel(
-    const long nthreads, const scalar_t* weight_data, const int* indices_data,
-    const uint16 nInputPlane, const uint16 nOutputPlane,
-    const uint8 nOrientation, const uint8 nRotation, const uint16 nEntry,
-    scalar_t* output_data) {
+    const int nthreads, const scalar_t* weight_data, const int* indices_data,
+    const int nInputPlane, const int nOutputPlane, const int nOrientation,
+    const int nRotation, const int nEntry, scalar_t* output_data) {
   CUDA_1D_KERNEL_LOOP(n, nthreads) {
     int l = n % nEntry;
     int j = (n / nEntry) % nInputPlane;
@@ -23,7 +22,7 @@ __global__ void ARF_forward_cuda_kernel(
     int k;
     scalar_t val = *(weight_data + n);
     for (k = 0; k < nRotation; k++) {
-      uint16 index = (uint16)(*(indices_data + l * nRotation + k)) - 1;
+      int index = (int)(*(indices_data + l * nRotation + k)) - 1;
       scalar_t* target = output_data + i * (nRotation * nInputPlane * nEntry) +
                          k * (nInputPlane * nEntry) + j * (nEntry) + index;
       *target = val;
@@ -45,7 +44,7 @@ __global__ void ARF_backward_cuda_kernel(
     scalar_t* val = weight_data + n;
     *val = 0;
     for (k = 0; k < nRotation; k++) {
-      int index = (uint16)(*(indices_data + l * nRotation + k)) - 1;
+      int index = (int)(*(indices_data + l * nRotation + k)) - 1;
       scalar_t target =
           *(gradWeight_data + i * (nRotation * nInputPlane * nEntry) +
             k * (nInputPlane * nEntry) + j * (nEntry) + index);
